Changing data types in db schema, entity classes and dtos. Correcting tests based on new data types. Creating connection to mysql database

+++ src/main/java/pl/myproject/car_rental_api/dto/CarDTO.java
@@ -17,7 +17,7 @@ public class CarDTO {
     private LocalDate lastServiceDate;
     private int mileage;
     private LocalDate insuranceExpiryDate;
-    private double rentalPricePerDay;
-    private double basePrice;
+    private int rentalPricePerDay;
+    private int basePrice;
     private ModelDTO model;
 }
